+++ CHANGELOG.md
@@ -18,7 +18,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/), and t
 
 ### 👷 Changed
 
-\-
+- Change useForm hook
+  - Add support for dropdown fields
 
 ### 📆 Deprecated